[tests] add tests with time
